Use re.search for body filter
re.search matches anywhere in the body unlike re.match which requires
the match to be positive from the beginning of the string. Using
re.search for request body match is more flexible and also allows for
correct matching in a multi-line strings - which the body of the
requests most likely going to be.